added dynamic refereshing display base

+++ src/objects/Vector2/Vector2.cpp
@@ -3,7 +3,7 @@
 namespace Snake {
 namespace Objects {
 
-Vector2::Vector2(int x, int y) : x(x), y(y) {}
+Vector2::Vector2(float x, float y) : x(x), y(y) {}
 
 Vector2 Vector2::Zero = Vector2(0, 0);
